promoted importance of ServiceModule
Deleted CoordinatedService and associated tests as it has been fully
deprecated by CoordinatedService2.

`ServiceModule` has been promoted to its own file, and documentation has
been reorganized to make sure it is easy to use. Provisioning a
`ServiceManager` will now a log a warning if a service is multibound in
the application, as they really should be installed via
`ServiceModule`s.
